Refactor ETL project structure and add tests (fixes #680)
This turns the Glean Dictionary into more of a standard python
package, like lookml-generator. Also add a single test for the Looker
ETL (will add more in the future, but this PR is big enough as it is)
