+++ .prettierignore
@@ -1,3 +1,4 @@
+.circleci
 public/build/bundle.js
 public/build/bundle.css
 public/data/*
