+++ netlify.toml
@@ -1,7 +1,8 @@
 [build]
 command = """
     set -e
-    scripts/build-glean-metadata.py
+    pip install -e .
+    build-glean-metadata
     npm run build
     if [ "$STORYBOOK" ]; then
       npm run build-storybook
