+++ pyproject.toml
@@ -11,3 +11,7 @@ multi_line_output = 3
 include_trailing_comma = "True"
 force_grid_wrap = 0
 use_parentheses = "True"
+
+[tool.pytest.ini_options]
+testpaths = ['etl', 'etl_tests']
+addopts = '--black --flake8 --isort'
