+++ requirements.in
@@ -1,9 +1,12 @@
 autoflake==1.4
 black==21.11b1
-flake8==4.0.1
-flake8-black==0.2.3
 furl==2.1.3
 isort==5.10.1
+pip-tools==6.4.0
+pytest==6.2.5
+pytest-black==0.3.12
+pytest-flake8==1.0.7
+pytest-isort==2.0.0
 requests==2.26.0
 stringcase==1.2.0
 PyYAML==6.0
