+++ scripts/build-glean-metadata
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+cd "$(dirname "$0")/.."
+
+if [ -e venv ]; then
+  . venv/bin/activate
+fi
+
+exec python3 -c "from etl.__main__ import main; main()" "$@"
