+++ scripts/lint-fix
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+cd "$(dirname "$0")/.."
+
+if [ -e venv ]; then
+  . venv/bin/activate
+fi
+
+autoflake -r -i --remove-all-unused-imports etl etl_tests
+black .
+isort .
